chore: Rename files to match Airbnb style guide
