+++ src/components/FancyRectangle.tsx
